requirements.txt
streamlit>=1.30.0
openai>=1.10.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
python-docx>=1.0.0
